Initial Commit
I've been hacking for a while, but forgot to commit.
Ok, so, at this point I think I've got tests for intersection between
rays and triangles and I've built a cube and can fire a ray at it.

It currently doesn't handle failure very well, and I don't know that
it's ever right.
I'm about to figure that out though, now that I've committed this.