Swapped (x, y) in Camera Ray Computation
So, I was using range ((0,0), (wres, hres)) to compute all the (x,y)
combinations.

The problem, though, was that range computes, sensibly, all ys for each
x.

This lead to [(0,0), (0,1), (0,2)..], which makes sense, but in this
case isn't what I wanted.

So, I instead computed the (0,0) - (hres, wres) range, then swapped them
all.
This gave me [(0,0), (1,0), (2, 0)..], which is what I wanted.