Various Changes
I'm just playing around, mostly.
So, the "approxZero" function was annoying me.

The issue, really, was that linearSolve threw an error on singular
matrix, so I had to figure that out first.

I tried using rank == 3 instead of the det, which returns a double, but
that was actually slower.

(luSolve . luPacked) is almost the same as linearSolve.
The difference is that it doesn't error out.

So, I just solve it, and if n is infinite, then it didn't hit.

I think this is still slower, though.

Also, once approxZero was out, only one place was using on2.
So, I figured that it wasn't necessary.

I replaced it with applicative functions for lists.
It's a few more characters, but may hopefully be more clear rather than
less.