Change Intersect To Return N
It used to return the intersection point.
Now, though, it returns the N along the ray where intersection occurs.

This is good, since if you don't care about where it collides, it does
less (potentially)

Also, though, it gives more info, in that I can now tell which two
things are closer to the viewer.
From n I can trivially get the actual point with calc_ray.

So this is good.

I also realized that up until now I haven't been checking that n was
positive (which is hard with the collision point, but trivial with n).
So, a plane which was "behind" the viewer was marked as a positive
collision.

While I was here I also factored a pattern out.
I now have all_of [(a -> Bool)] -> a -> Bool.
It returns true if all things are true about a, and False otherwise.