Made Cube's intersection More Correct
It used to return the first thing it collided with.
That was more efficient, but not right.

Now it returns the first thing the ray would have collided with, rather
than the first item in the arbitrary list that was a collision.

So, it takes longer now, since it actually has to compute collision with
each item and then find the smallest n, but it's more correct.