Cleaned Up Triangle Ray Intersection
So, it was figuring out where the plane the triangle lay on collided with
the ray, then figuring out where on the triangle that ray hit.

So, then I realized that the collision between the ray and the plane
returned 3 numbers, and I was only using one.
The other two were the ones I was solving for later.
Poo.

So, I've cut that out so that I now solve once and use those numbers.
It's taken my simple scene from 1m22s on my system to 33s, which is
nice.

Still not super quick, but better.