Fixed Camera Generation
I was dumb.
So, my math uses the Cross Product to figure out the axis the height and
width of the resultant image.

I forgot the step, though, that makes them a unit axis again, I was just
assuming they would be.

So, I was sometimes ending up with an axis of [0,1,0], but when I
"turned" the camera it'd be [0,2,0] or whatever.

That made things get weird...