Publishize
When I came back to this repo on a different computer 4 years later, I didn't
know what I needed to install to get it to run.
When I did install things, it installed the newest versions, which were
different than what the library was originally built with.

So... I have a cabal file now.