N Parallel Method
parGenerateImage now works with N threads, where N is the number of threads the
executable is run with.

As an optimization, if N is 1, it just does a straight generateImage.

I've found that this method seems to be slightly faster than the last, even
when N is 2.
Which is good, because even though I have 4 cores, N=2 is fastest.

Tradeoffs...