Add Color to Mesh
I now attach a color (or whatever) to a mesh, and I can my ray collision
returns the distance of the collision along with the value we plucked from
there.

That will be the first step in the lighting I'm planning on working on.

For now I've added another mesh to our scene so I can see the two colours.
I've also changed the pixel generation to take both distance and colour into
account.

On an unrelated note, I think my camera might be upside down...
I'll look into that in a bit maybe?