New Test Image
I moved the scene around so it looks more interesting.
In doing this I realized that the image was upside down because the it counted
the 0th line as the top of the image, but my 3D space assumed positive y was
"up".

So, I just flip the rows around now.