Added Shadows!
Ok, so, I'm now "actually" ray tracing, and not just "distance casting"

When my ray hits a point, I find the point it hits and then see if that point
can see any of my lights.

If it can, then I draw it the colour of the mesh, and if it's not, then I draw
it 1/5th that colour.

This gives me shadows.

I want to finesse the lights a bit so they have an intensity that dies off to
make the shapes more dynamic.
I also want to colour the lights and have that blend with the colour of the
object.

That is all bonus, though.
This traces rays.
Woot.

Oh, also, there are some points that aren't lighting right.
I assume these are floating point issues, but I should probably make sure.