Mix Light and Object Colour
Ok, so, the light on a spot adds up, and then the colour of the object affects
that.

For now I'm adding the light by just adding and capping at 255, so a full
intensity R and G will yield a full intensity Y, but white light just washes
over everything.

I hope that will change with light intensity stuff.

Also, I had to change the colour of the blue light, because blue light doesn't
show up on a yellow object.