Clear Up Lighting Artifacts
The issue was floating point stuff, which I figured.
This time it was that some of the points on the flat surface were computed not
to be at 0.0, but at -4E-18 or whatever, which meant it was under the surface,
and so the ray collided with the fact itself when seeing if it's illuminated.

I don't know if I love this solution, but basically I'm just rounding 3D points
to the nearest millionth, and a that point the smoothed out shape is no longer
in its own shade.

In the debugging process I made camera implement Show, which most things were
doing anyway, so I left it in.